+++ javascript/node/selenium-webdriver/test/window_test.js
@@ -15,8 +15,6 @@
 
 'use strict';
 
-var assert = require('assert');
-
 var test = require('../lib/test');