+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -1,6 +1,6 @@
 ## v2.34.0-dev
 
-* Added the `selenium-webdriver/testing/asserts` module. This module
+* Added the `selenium-webdriver/testing/assert` module. This module
     simplifies writing assertions against promised values (see
     example in module documentation).
 * Added the `webdriver.Capabilities` class.